[Package] 게이트웨이에 필터 패키지 생성
