+++ msa.gateway/src/main/java/com/msa/gateway/filter/CustomPostFilter.java
@@ -1,28 +1,27 @@
-package com.msa.gateway;
+package com.msa.gateway.filter;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.cloud.gateway.filter.GatewayFilterChain;
 import org.springframework.cloud.gateway.filter.GlobalFilter;
 import org.springframework.core.Ordered;
-import org.springframework.http.server.reactive.ServerHttpRequest;
+import org.springframework.http.server.reactive.ServerHttpResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ServerWebExchange;
 import reactor.core.publisher.Mono;
 
-@Slf4j
 @Component
-public class CustomPreFilter implements GlobalFilter, Ordered {
-
-
+public class CustomPostFilter implements GlobalFilter, Ordered {
     @Override
     public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
-        ServerHttpRequest response = exchange.getRequest();
-        log.info("Pre Filter: Request URI is {}", response.getURI());
-        return chain.filter(exchange);
+        return chain.filter(exchange).then(Mono.fromRunnable(() -> {
+            ServerHttpResponse response = exchange.getResponse();
+            /*
+            Add any custom logic here
+             */
+        }));
     }
 
     @Override
     public int getOrder() {
-        return  Ordered.HIGHEST_PRECEDENCE;
+        return Ordered.LOWEST_PRECEDENCE;
     }
 }
