+++ msa.gateway/src/main/java/com/msa/gateway/filter/CustomPreFilter.java
@@ -1,27 +1,28 @@
-package com.msa.gateway;
+package com.msa.gateway.filter;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.cloud.gateway.filter.GatewayFilterChain;
 import org.springframework.cloud.gateway.filter.GlobalFilter;
 import org.springframework.core.Ordered;
-import org.springframework.http.server.reactive.ServerHttpResponse;
+import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ServerWebExchange;
 import reactor.core.publisher.Mono;
 
+@Slf4j
 @Component
-public class CustomPostFilter implements GlobalFilter, Ordered {
+public class CustomPreFilter implements GlobalFilter, Ordered {
+
+
     @Override
     public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
-        return chain.filter(exchange).then(Mono.fromRunnable(() -> {
-            ServerHttpResponse response = exchange.getResponse();
-            /*
-            Add any custom logic here
-             */
-        }));
+        ServerHttpRequest response = exchange.getRequest();
+        log.info("Pre Filter: Request URI is {}", response.getURI());
+        return chain.filter(exchange);
     }
 
     @Override
     public int getOrder() {
-        return Ordered.LOWEST_PRECEDENCE;
+        return  Ordered.HIGHEST_PRECEDENCE;
     }
 }
